winkelmandje werkt helemaal, enkel order opslaan lukt niet wegens timestamp

+++ src/main/java/com/realdolmen/shopr/domain/Order.java
@@ -5,11 +5,14 @@ import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.List;
 
-// TODO: mappedBy probleem op stackoverflow opzoeken
 // TODO: growl message met bevestiging van bestelling bij "add-item-to-shopping-cart" button
 // TODO: confirm dialog bij delete button -> backend moet nog message meegeven voor de growl message
-// TODO: flow bij proceed to checkout button click
 // TODO: waarom geen primefaces icons in commandbuttons?
+// TODO: op artikeldetail: aantal default op 1 zetten
+// TODO: bij overzicht winkelmandje: delete knop geeft altijd 0 mee als id
+// TODO: bij overzicht winkelmandje: edit knop implementeren
+// TODO:
+// TODO:
 
 @Entity
 public class Order {
@@ -68,7 +71,7 @@ public class Order {
 
     public void setPrice() {
         if (this.getOrderLines() != null && this.getOrderLines().size() > 0) {
-            int tempPrice = 0;
+            double tempPrice = 0;
             for (OrderLine orderLine : this.getOrderLines()) {
                 tempPrice += orderLine.getSubTotal();
             }
